+++ CHANGELOG.md
@@ -2,9 +2,13 @@
 ## 1.02.0
   **ТРЕБУЕТСЯ ПРОШИТЬ РАМКУ**: версией 1.02
 
+  **ФОРМАТ ФАЙЛА prompts.json ИЗМЕНЁН** на prompts.yaml
+
 - Add threshold for providers.  [#1](https://github.com/maxifly/ImageServer/issues/1)
-- Add sleep mode.  [#6](https://github.com/maxifly/ImageServer/issues/6)
+- Add template prompts [#2](https://github.com/maxifly/ImageServer/issues/2)
 - Add metrics.  [#3](https://github.com/maxifly/ImageServer/issues/3)
+- Add sleep mode.  [#6](https://github.com/maxifly/ImageServer/issues/6)
+
 
 
 ## 1.1